[test] refactoring
