+++ src/test/java/com/ibm/stocator/fs/swift2d/systemtests/CollisionTest.java
@@ -26,8 +26,9 @@ import org.junit.Assume;
 import org.junit.Test;
 
 import com.ibm.stocator.fs.common.FileSystemTestUtils;
+import com.ibm.stocator.fs.common.BaseTest;
 
-public class CollisionTest extends SwiftBaseTest {
+public class CollisionTest extends BaseTest {
 
   protected byte[] data = FileSystemTestUtils.generateDataset(getBlockSize() * 2, 0, 255);
   protected byte[] smData = FileSystemTestUtils.generateDataset(getBlockSize(), 0, 255);
